optimize the get where condition
