+++ rm-datasource/src/main/java/io/seata/rm/datasource/sql/druid/BaseRecognizer.java
@@ -15,14 +15,7 @@
  */
 package io.seata.rm.datasource.sql.druid;
 
-import com.alibaba.druid.sql.ast.expr.SQLVariantRefExpr;
-import com.alibaba.druid.sql.dialect.mysql.visitor.MySqlOutputVisitor;
-import com.alibaba.druid.sql.dialect.oracle.visitor.OracleOutputVisitor;
-import io.seata.rm.datasource.ParametersHolder;
 import io.seata.rm.datasource.sql.SQLRecognizer;
-
-import java.util.ArrayList;
-import java.util.List;
 
 /**
  * The type Base recognizer.
@@ -60,47 +53,5 @@ public abstract class BaseRecognizer implements SQLRecognizer {
     @Override
     public String getOriginalSQL() {
         return originalSQL;
-    }
-
-    public MySqlOutputVisitor createMySqlOutputVisitor(final ParametersHolder parametersHolder, final ArrayList<List<Object>> paramAppenderList, final StringBuilder sb) {
-        MySqlOutputVisitor visitor = new MySqlOutputVisitor(sb) {
-
-            @Override
-            public boolean visit(SQLVariantRefExpr x) {
-                if ("?".equals(x.getName())) {
-                    ArrayList<Object> oneParamValues = parametersHolder.getParameters()[x.getIndex()];
-                    if (paramAppenderList.size() == 0) {
-                        oneParamValues.stream().forEach(t -> paramAppenderList.add(new ArrayList<>()));
-                    }
-                    for (int i = 0; i < oneParamValues.size(); i++) {
-                        paramAppenderList.get(i).add(oneParamValues.get(i));
-                    }
-
-                }
-                return super.visit(x);
-            }
-        };
-        return visitor;
-    }
-
-    public OracleOutputVisitor createOracleOutputVisitor(final ParametersHolder parametersHolder, final ArrayList<List<Object>> paramAppenderList, final StringBuilder sb) {
-        OracleOutputVisitor visitor = new OracleOutputVisitor(sb) {
-
-            @Override
-            public boolean visit(SQLVariantRefExpr x) {
-                if ("?".equals(x.getName())) {
-                    ArrayList<Object> oneParamValues = parametersHolder.getParameters()[x.getIndex()];
-                    if (paramAppenderList.size() == 0) {
-                        oneParamValues.stream().forEach(t -> paramAppenderList.add(new ArrayList<>()));
-                    }
-                    for (int i = 0; i < oneParamValues.size(); i++) {
-                        paramAppenderList.get(i).add(oneParamValues.get(i));
-                    }
-
-                }
-                return super.visit(x);
-            }
-        };
-        return visitor;
     }
 }
