+++ rm-datasource/src/main/java/io/seata/rm/datasource/sql/druid/oracle/BaseOracleRecognizer.java
@@ -1,0 +1,102 @@
+/*
+ *  Copyright 1999-2019 Seata.io Group.
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+package io.seata.rm.datasource.sql.druid.oracle;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import com.alibaba.druid.sql.ast.SQLExpr;
+import com.alibaba.druid.sql.ast.expr.SQLBetweenExpr;
+import com.alibaba.druid.sql.ast.expr.SQLBinaryOpExpr;
+import com.alibaba.druid.sql.ast.expr.SQLInListExpr;
+import com.alibaba.druid.sql.ast.expr.SQLVariantRefExpr;
+import com.alibaba.druid.sql.dialect.oracle.visitor.OracleOutputVisitor;
+
+import io.seata.rm.datasource.ParametersHolder;
+import io.seata.rm.datasource.sql.druid.BaseRecognizer;
+
+/**
+ * @author will
+ * @date 2019/9/26
+ */
+public abstract class BaseOracleRecognizer extends BaseRecognizer {
+
+    /**
+     * Instantiates a new oracle base recognizer
+     * @param originalSql the original sql
+     */
+    public BaseOracleRecognizer(String originalSql){
+        super(originalSql);
+    }
+
+    public OracleOutputVisitor createOutputVisitor(final ParametersHolder parametersHolder, final ArrayList<List<Object>> paramAppenderList, final StringBuilder sb) {
+        OracleOutputVisitor visitor = new OracleOutputVisitor(sb) {
+
+            @Override
+            public boolean visit(SQLVariantRefExpr x) {
+                if ("?".equals(x.getName())) {
+                    ArrayList<Object> oneParamValues = parametersHolder.getParameters()[x.getIndex()];
+                    if (paramAppenderList.size() == 0) {
+                        oneParamValues.stream().forEach(t -> paramAppenderList.add(new ArrayList<>()));
+                    }
+                    for (int i = 0; i < oneParamValues.size(); i++) {
+                        paramAppenderList.get(i).add(oneParamValues.get(i));
+                    }
+
+                }
+                return super.visit(x);
+            }
+        };
+        return visitor;
+    }
+
+    public String getWhereCondition(SQLExpr where, final ParametersHolder parametersHolder, final ArrayList<List<Object>> paramAppenderList) {
+        if (where == null) {
+            return "";
+        }
+        StringBuilder sb = new StringBuilder();
+        OracleOutputVisitor visitor = createOutputVisitor(parametersHolder, paramAppenderList, sb);
+        if (where instanceof SQLBinaryOpExpr) {
+            visitor.visit((SQLBinaryOpExpr) where);
+        } else if (where instanceof SQLInListExpr) {
+            visitor.visit((SQLInListExpr) where);
+        } else if (where instanceof SQLBetweenExpr) {
+            visitor.visit((SQLBetweenExpr) where);
+        } else {
+            throw new IllegalArgumentException("unexpected WHERE expr: " + where.getClass().getSimpleName());
+        }
+        return sb.toString();
+    }
+
+    public String getWhereCondition(SQLExpr where) {
+        if (where == null) {
+            return "";
+        }
+
+        StringBuilder sb = new StringBuilder();
+        OracleOutputVisitor visitor = new OracleOutputVisitor(sb);
+        if (where instanceof SQLBinaryOpExpr) {
+            visitor.visit((SQLBinaryOpExpr) where);
+        } else if (where instanceof SQLInListExpr) {
+            visitor.visit((SQLInListExpr) where);
+        } else if (where instanceof SQLBetweenExpr) {
+            visitor.visit((SQLBetweenExpr) where);
+        } else {
+            throw new IllegalArgumentException("unexpected WHERE expr: " + where.getClass().getSimpleName());
+        }
+        return sb.toString();
+    }
+}
